requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cloudpickle==3.0.0
Cython==0.29.37
decorator==5.1.1
dm-tree==0.1.8
etils==1.5.2
fasteners==0.19
flatbuffers==24.3.25
fsspec==2024.3.1
gast==0.5.4
gin-config==0.5.0
glfw==2.7.0
glog==0.3.1
grpcio==1.62.2
gym==0.23.0
gym-notices==0.0.8
h5py==3.11.0
idna==3.7
imageio==2.34.1
importlib_metadata==7.1.0
importlib_resources==6.4.0
keras==2.14.0
libclang==18.1.1
Markdown==3.6
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml-dtypes==0.2.0
mujoco-py==2.1.2.14
namex==0.0.8
numpy==1.26.4
oauthlib==3.2.2
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pillow==10.3.0
protobuf==4.25.3
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pygame==2.1.3
pyglib==0.1
Pygments==2.17.2
PyOpenGL==3.1.7
python-gflags==3.1.2
requests==2.31.0
requests-oauthlib==2.0.0
rich==13.7.1
rsa==4.9
scipy==1.13.0
six==1.16.0
tensorboard==2.14.1
tensorboard-data-server==0.7.2
tensorflow==2.14.0
tensorflow-estimator==2.14.0
tensorflow-probability==0.22.0
termcolor==2.4.0
tf-agents==0.18.0
tf-keras==2.14.1
typing_extensions==4.5.0
urllib3==2.2.1
Werkzeug==3.0.2
wrapt==1.14.1
zipp==3.18.1